--- a/itou/approvals/models.py
+++ b/itou/approvals/models.py
@@ -87,28 +87,20 @@ class CommonApprovalMixin(models.Model):
     def duration(self):
         return self.end_at - self.start_at
 
-    def _get_obj_remainder(self, obj):
-        """
-        Return the remaining time on an object with start_at and end_at dete fields
-        A.k.a an Approval, a Suspension or a Prolongation
-        """
-        return max(obj.end_at - timezone.localdate(), datetime.timedelta(0)) - max(
-            obj.start_at - timezone.localdate(), datetime.timedelta(0)
-        )
-
-    @cached_property
+    @property
     def remainder(self):
-        """
-        Return the remaining time of an Approval, we don't count future suspended periods.
-        """
-        result = self._get_obj_remainder(self)
-
-        if hasattr(self, "suspension_set"):
+        """Return the remaining time of an Approval."""
+        today = timezone.localdate()
+        zero = datetime.timedelta(0)
+        result = max(self.end_at - today, zero) - max(self.start_at - today, zero)
+        try:
+            suspended_until = self.suspended_until
+        except AttributeError:
             # PoleEmploiApprovals don't have suspensions
-            result -= sum(
-                (self._get_obj_remainder(suspension) for suspension in self.suspension_set.all()),
-                datetime.timedelta(0),
-            )
+            pass
+        else:
+            if suspended_until:
+                result -= max(suspended_until - today, zero)
         return result
 
     @property