**Carte Notion : https://www.notion.so/plateforme-inclusion/Suite-journ-e-dev-Approval-suspended_until-6455d136d38146a192c9bfb04c09e1eb**

### Performances

Méthode : Best of 3, selon l’onglet SQL de la DjDT (arrondi à 5 ms)

#### SIAE
Utilisateur : Un utilisateur d’une grosse SIAE (2700 candidatures)

| Nom | Master | PR | URL |
|-----|--------|----|-----|
| Candidatures NEW & PROCESSING | 180 ms | 180 ms | http://localhost:8000/apply/siae/list?states=new&states=processing |
| Candidatures NEW & PROCESSING Eligibilité validée | 173 ms | 171 ms | http://localhost:8000/apply/siae/list?states=new&states=processing&eligibility_validated=on&start_date=&end_date= |
| Eligibilité validée, suspendu | 440 ms | 290 ms | http://localhost:8000/apply/siae/list?eligibility_validated=on&pass_iae_suspended=on&start_date=&end_date= |
| Eligibilité validée, actif | 420 ms | 275 ms | http://localhost:8000/apply/siae/list?eligibility_validated=on&pass_iae_active=on&start_date=&end_date= |
| Gérer les PASS IAE | 110 ms | 110 ms | http://localhost:8000/approvals/list |
| Gérer les PASS IAE suspendus | 360 ms | 105 ms | http://localhost:8000/approvals/list?status_suspended=on |
| Gérer les PASS IAE valides | 600 ms | 105 ms | http://localhost:8000/approvals/list?status_valid=on |
| Gérer les PASS IAE futurs | 110 ms | 105 ms | http://localhost:8000/approvals/list?status_future=on |
| Gérer les fiches salarié | 225 ms | 225 ms | http://localhost:8000/employee_record/list?status=NEW |

#### Prescripteur

Utilisateur : Un prescripteur ayant orienté beaucoup de candidats (1350 candidatures)

| Nom | Master | PR | URL |
|-----|--------|----|-----|
| Candidatures | 235 ms | 210 ms | http://localhost:8000/apply/prescriber/list |
| Candidatures avec éligibilité validée | 260 ms | 225 ms | http://localhost:8000/apply/prescriber/list?start_date=&end_date=&eligibility_validated=on |