@francoisfreitag 2023-07-18
Le champ aide sur quelques vues. Pour la plupart, il n’apporte pas de changement de performances.
Comme il est assez facile à maintenir, je pense qu’on peut l’intégrer, mais ravi d’en discuter :smiley: