@francoisfreitag 2023-07-26
Après un rebase sur master, les chiffres de performances n’ont pas changé significativement par rapport au tableau en description.

La vue qui en bénéficie le plus est “[Gérer les PASS IAE](http://localhost:8000/approvals/list)”. Elle semble avoir été utilisée environ 300K fois les trois derniers mois, soit environ 3 000 fois par jour (voir https://itou-inclusion.slack.com/archives/C0412CTV63D/p1690363912372609).

Les requêtes les plus lentes sur sur cette vue (sur les 14 derniers jours) n’utilisent pas les filtres qui bénéficient d’une amélioration de performance. Donc l’amélioration de perfs me semble assez maginale.

Sauf avis contraire, je pense clore cette PR dans la semaine et ne pas intégrer la modification.
- [ ] J’ajouterai une note dans le modèle approvals pour diriger vers cette expérimentation infructueuse.